# spring/classes/openapi/api-spec.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Abacus WEB Sample",
    "description": "Abacus WEB Sample Project",
    "license": {
      "name": "Apache 2.0",
      "url": "http://abacus.co.kr"
    },
    "version": "v1.0"
  },
  "servers": [
    {
      "url": "http://192.168.2.142:8080/api/",
      "description": "dev"
    }
  ],
...
  "paths": {
...
    "/serviceServer/updateServiceServer": {
      "post": {
        "tags": [
          "서비스 등록 - 서비스 서버 등록"
        ],
        "summary": "서비스 서버 수정",
        "description": "서비스 서버 수정",
        "operationId": "updateServiceServer",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "svcSvrCd": {
                    "type": "string",
                    "example": "TEST"
                  },
                  "svcSvrNm": {
                    "type": "string",
                    "example": "테스느 서비스 서버"
                  },
                  "svcSvrType": {
                    "type": "string",
                    "example": "asn"
                  },
                  "expireDuration": {
                    "type": "string",
                    "example": 999
                  },
                  "expireUnit": {
                    "type": "string",
                    "example": "YR"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/serviceServer/retrieveServiceServer": {
      "post": {
        "tags": [
          "서비스 등록 - 서비스 서버 등록"
        ],
        "summary": "서비스 서버 상세 조회",
        "description": "서비스 서버 상세 조회",
        "operationId": "retrieveServiceServer",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "svcSvrCd": {
                    "type": "string",
                    "example": "TEST"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComResponseSvcServer"
                }
              }
            }
          }
        }
      }
    },
    "/serviceServer/retrieveServiceServerList": {
      "post": {
        "tags": [
          "서비스 등록 - 서비스 서버 등록"
        ],
        "summary": "서비스 서버 목록 조회",
        "description": "서비스 서버 목록 조회",
        "operationId": "retrieveServiceServerList",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "poa": {
                    "type": "string",
                    "example": ""
                  },
                  "svcCd": {
                    "type": "string",
                    "example": ""
                  },
                  "currentPage": {
                    "type": "string",
                    "example": 1
                  },
                  "limit": {
                    "type": "string",
                    "example": 10
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComResponseSvcServerList"
                }
              }
            }
          }
        }
      }
    },
    "/serviceServer/poa/updatePoa": {
      "post": {
        "tags": [
          "서비스 서버 등록"
        ],
        "summary": "POA 수정",
        "description": "POA 수정",
        "operationId": "updatePoa",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "entityId": {
                    "type": "string",
                    "example": ""
                  },
                  "resourceNm": {
                    "type": "string",
                    "example": ""
                  },
                  "poa": {
                    "type": "string",
                    "example": "http://192.168.2.131:8280"
                  },
                  "contentType": {
                    "type": "string",
                    "example": "JSON"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/serviceServer/poa/retrievePoa": {
      "post": {
        "tags": [
          "서비스 서버 등록"
        ],
        "summary": "POA 상세 조회",
        "description": "POA 상세 조회",
        "operationId": "retrievePoa",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "entityId": {
                    "type": "string",
                    "example": "ASN_CSE-S-bfc7601bb7-V040"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComResponsePoaHistory"
                }
              }
            }
          }
        }
      }
    },
    "/serviceServer/poa/retrievePoaHistoryList": {
      "post": {
        "tags": [
          "서비스 서버 등록"
        ],
        "summary": "POA 이력 목록 조회",
        "description": "POA 이력 목록 조회",
        "operationId": "retrievePoaHistoryList",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "entityId": {
                    "type": "string",
                    "example": "ASN_CSE-S-bfc7601bb7-V040"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComResponsePoaHistoryList"
                }
              }
            }
          }
        }
      }
    },
...
    "/serviceServer/poa/deletePoa": {
      "post": {
        "tags": [
          "서비스 서버 등록"
        ],
        "summary": "POA 삭제",
        "description": "POA 삭제",
        "operationId": "deletePoa",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "entityId": {
                    "type": "string",
                    "example": ""
                  },
                  "resourceNm": {
                    "type": "string",
                    "example": ""
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/serviceServer/poa/createPoa": {
      "post": {
        "tags": [
          "서비스 서버 등록"
        ],
        "summary": "POA 등록",
        "description": "POA 등록",
        "operationId": "createPoa",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "svcSvrCd": {
                    "type": "string",
                    "example": 4
                  },
                  "entityId": {
                    "type": "string",
                    "example": "ASN_CSE-S-bfc7601bb7-V040"
                  },
                  "resourceNm": {
                    "type": "string",
                    "example": "ASN_CSE-S-bfc7601bb7-V040"
                  },
                  "poa": {
                    "type": "string",
                    "example": "http://192.168.2.131:8280"
                  },
                  "contentType": {
                    "type": "string",
                    "example": "JSON"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/serviceServer/downloadSvcServerList": {
      "post": {
        "tags": [
          "서비스 등록 - 서비스 서버 등록"
        ],
        "summary": "서비스 서버 엑셀 다운로드",
        "description": "서비스 서버 엑셀 다운로드",
        "operationId": "downloadSvcServerList",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "svcSvrNm": {
                    "type": "string",
                    "example": "TEST"
                  },
                  "svcCd": {
                    "type": "string",
                    "example": "TEST"
                  },
                  "startDate": {
                    "type": "string",
                    "example": "2025-12-01"
                  },
                  "endDate": {
                    "type": "string",
                    "example": "2025-12-01"
                  },
                  "sidx": {
                    "type": "string",
                    "example": "creDt"
                  },
                  "sord": {
                    "type": "string",
                    "example": "desc"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/serviceServer/deleteServiceServer": {
      "post": {
        "tags": [
          "서비스 등록 - 서비스 서버 등록"
        ],
        "summary": "서비스 서버 삭제",
        "description": "서비스 서버 삭제",
        "operationId": "deleteServiceServer",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "svcSvrCd": {
                    "type": "string",
                    "example": "TEST"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/serviceServer/createServiceServer": {
      "post": {
        "tags": [
          "서비스 등록 - 서비스 서버 등록"
        ],
        "summary": "서비스 서버 등록",
        "description": "서비스 서버 등록",
        "operationId": "createServiceServer",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "svcSvrNm": {
                    "type": "string",
                    "example": "테스트 서비스 서버"
                  },
                  "svcSvrType": {
                    "type": "string",
                    "example": "asn"
                  },
                  "svcCd": {
                    "type": "string",
                    "example": "TEST"
                  },
                  "expireDuration": {
                    "type": "string",
                    "example": 999
                  },
                  "expireUnit": {
                    "type": "string",
                    "example": "YR"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ComResponseSvcServer": {
        "type": "object",
        "properties": {
          "httpStatusCode": {
            "type": "integer",
            "format": "int32"
          },
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "detailMessage": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "body": {
            "$ref": "#/components/schemas/SvcServer"
          },
          "errorMessage": {
            "type": "string"
          }
        }
      },
...
      "ComResponseSvcServerList": {
        "type": "object",
        "properties": {
          "httpStatusCode": {
            "type": "integer",
            "format": "int32"
          },
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "detailMessage": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "body": {
            "$ref": "#/components/schemas/SvcServerList"
          },
          "errorMessage": {
            "type": "string"
          }
        }
      },
...
      "ComResponsePoaHistory": {
        "type": "object",
        "properties": {
          "httpStatusCode": {
            "type": "integer",
            "format": "int32"
          },
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "detailMessage": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "body": {
            "$ref": "#/components/schemas/PoaHistory"
          },
          "errorMessage": {
            "type": "string"
          }
        }
      },
...
    }
  }
}