# spring/generated-sources/openapi/src/main/java/com/openapi/generator/aicc/api/ServiceServerApi.java
/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.18.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.openapi.generator.aicc.api;

import com.openapi.generator.aicc.dto.ComResponsePoaHistory;
import com.openapi.generator.aicc.dto.ComResponsePoaHistoryList;
import com.openapi.generator.aicc.dto.ComResponseString;
import com.openapi.generator.aicc.dto.ComResponseSvcServer;
import com.openapi.generator.aicc.dto.ComResponseSvcServerList;
import com.openapi.generator.aicc.dto.CreatePoaRequest;
import com.openapi.generator.aicc.dto.CreateServiceServerRequest;
import com.openapi.generator.aicc.dto.DeletePoaRequest;
import com.openapi.generator.aicc.dto.DownloadSvcServerListRequest;
import com.openapi.generator.aicc.dto.HealthCheckRequest;
import com.openapi.generator.aicc.dto.RetrievePoaRequest;
import com.openapi.generator.aicc.dto.RetrieveServiceServerListRequest;
import com.openapi.generator.aicc.dto.RetrieveServiceServerRequest;
import com.openapi.generator.aicc.dto.UpdatePoaRequest;
import com.openapi.generator.aicc.dto.UpdateServiceServerRequest;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.ExampleObject;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-01-19T16:30:08.398412+09:00[Asia/Seoul]", comments = "Generator version: 7.18.0")
@Validated
@Tag(name = "서비스 서버 등록", description = "POA")
public interface ServiceServerApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    String PATH_CREATE_POA = "/serviceServer/poa/createPoa";
    /**
     * POST /serviceServer/poa/createPoa : POA 등록
     * POA 등록
     *
     * @param createPoaRequest  (required)
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "createPoa",
        summary = "POA 등록",
        description = "POA 등록",
        tags = { "서비스 서버 등록" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = ServiceServerApi.PATH_CREATE_POA,
        consumes = { "application/json" }
    )
    default ResponseEntity<Void> createPoa(
        @Parameter(name = "CreatePoaRequest", description = "", required = true) @Valid @RequestBody CreatePoaRequest createPoaRequest
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_CREATE_SERVICE_SERVER = "/serviceServer/createServiceServer";
    /**
     * POST /serviceServer/createServiceServer : 서비스 서버 등록
     * 서비스 서버 등록
     *
     * @param createServiceServerRequest  (required)
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "createServiceServer",
        summary = "서비스 서버 등록",
        description = "서비스 서버 등록",
        tags = { "서비스 등록 - 서비스 서버 등록" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = ServiceServerApi.PATH_CREATE_SERVICE_SERVER,
        consumes = { "application/json" }
    )
    default ResponseEntity<Void> createServiceServer(
        @Parameter(name = "CreateServiceServerRequest", description = "", required = true) @Valid @RequestBody CreateServiceServerRequest createServiceServerRequest
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_DELETE_POA = "/serviceServer/poa/deletePoa";
    /**
     * POST /serviceServer/poa/deletePoa : POA 삭제
     * POA 삭제
     *
     * @param deletePoaRequest  (required)
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "deletePoa",
        summary = "POA 삭제",
        description = "POA 삭제",
        tags = { "서비스 서버 등록" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = ServiceServerApi.PATH_DELETE_POA,
        consumes = { "application/json" }
    )
    default ResponseEntity<Void> deletePoa(
        @Parameter(name = "DeletePoaRequest", description = "", required = true) @Valid @RequestBody DeletePoaRequest deletePoaRequest
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_DELETE_SERVICE_SERVER = "/serviceServer/deleteServiceServer";
    /**
     * POST /serviceServer/deleteServiceServer : 서비스 서버 삭제
     * 서비스 서버 삭제
     *
     * @param retrieveServiceServerRequest  (required)
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "deleteServiceServer",
        summary = "서비스 서버 삭제",
        description = "서비스 서버 삭제",
        tags = { "서비스 등록 - 서비스 서버 등록" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = ServiceServerApi.PATH_DELETE_SERVICE_SERVER,
        consumes = { "application/json" }
    )
    default ResponseEntity<Void> deleteServiceServer(
        @Parameter(name = "RetrieveServiceServerRequest", description = "", required = true) @Valid @RequestBody RetrieveServiceServerRequest retrieveServiceServerRequest
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_DOWNLOAD_SVC_SERVER_LIST = "/serviceServer/downloadSvcServerList";
    /**
     * POST /serviceServer/downloadSvcServerList : 서비스 서버 엑셀 다운로드
     * 서비스 서버 엑셀 다운로드
     *
     * @param downloadSvcServerListRequest  (required)
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "downloadSvcServerList",
        summary = "서비스 서버 엑셀 다운로드",
        description = "서비스 서버 엑셀 다운로드",
        tags = { "서비스 등록 - 서비스 서버 등록" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = ServiceServerApi.PATH_DOWNLOAD_SVC_SERVER_LIST,
        consumes = { "application/json" }
    )
    default ResponseEntity<Void> downloadSvcServerList(
        @Parameter(name = "DownloadSvcServerListRequest", description = "", required = true) @Valid @RequestBody DownloadSvcServerListRequest downloadSvcServerListRequest
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_HEALTH_CHECK = "/serviceServer/poa/health-check";
    /**
     * POST /serviceServer/poa/health-check : 서비스서버 통신상태 확인
     * 서비스서버 통신상태 확인
     *
     * @param healthCheckRequest  (required)
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "healthCheck",
        summary = "서비스서버 통신상태 확인",
        description = "서비스서버 통신상태 확인",
        tags = { "서비스 서버 등록" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ComResponseString.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = ServiceServerApi.PATH_HEALTH_CHECK,
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    default ResponseEntity<ComResponseString> healthCheck(
        @Parameter(name = "HealthCheckRequest", description = "", required = true) @Valid @RequestBody HealthCheckRequest healthCheckRequest
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"code\" : \"code\", \"errorMessage\" : \"errorMessage\", \"detailMessage\" : \"detailMessage\", \"message\" : \"message\", \"body\" : \"body\", \"httpStatusCode\" : 0, \"status\" : \"status\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_RETRIEVE_POA = "/serviceServer/poa/retrievePoa";
    /**
     * POST /serviceServer/poa/retrievePoa : POA 상세 조회
     * POA 상세 조회
     *
     * @param retrievePoaRequest  (required)
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "retrievePoa",
        summary = "POA 상세 조회",
        description = "POA 상세 조회",
        tags = { "서비스 서버 등록" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ComResponsePoaHistory.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = ServiceServerApi.PATH_RETRIEVE_POA,
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    default ResponseEntity<ComResponsePoaHistory> retrievePoa(
        @Parameter(name = "RetrievePoaRequest", description = "", required = true) @Valid @RequestBody RetrievePoaRequest retrievePoaRequest
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"code\" : \"code\", \"errorMessage\" : \"errorMessage\", \"detailMessage\" : \"detailMessage\", \"message\" : \"message\", \"body\" : { \"poa\" : \"poa\", \"resourceNm\" : \"resourceNm\", \"origin\" : \"origin\", \"entityId\" : \"entityId\", \"svcCd\" : \"svcCd\", \"updUsr\" : \"updUsr\", \"creDt\" : \"creDt\", \"rolNm\" : \"rolNm\", \"poaStatus\" : \"poaStatus\", \"creType\" : \"creType\", \"useYn\" : \"useYn\", \"creUsr\" : \"creUsr\", \"updDt\" : \"updDt\", \"contentType\" : \"contentType\", \"seq\" : \"seq\" }, \"httpStatusCode\" : 0, \"status\" : \"status\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_RETRIEVE_POA_HISTORY_LIST = "/serviceServer/poa/retrievePoaHistoryList";
    /**
     * POST /serviceServer/poa/retrievePoaHistoryList : POA 이력 목록 조회
     * POA 이력 목록 조회
     *
     * @param retrievePoaRequest  (required)
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "retrievePoaHistoryList",
        summary = "POA 이력 목록 조회",
        description = "POA 이력 목록 조회",
        tags = { "서비스 서버 등록" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ComResponsePoaHistoryList.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = ServiceServerApi.PATH_RETRIEVE_POA_HISTORY_LIST,
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    default ResponseEntity<ComResponsePoaHistoryList> retrievePoaHistoryList(
        @Parameter(name = "RetrievePoaRequest", description = "", required = true) @Valid @RequestBody RetrievePoaRequest retrievePoaRequest
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"code\" : \"code\", \"errorMessage\" : \"errorMessage\", \"detailMessage\" : \"detailMessage\", \"message\" : \"message\", \"body\" : { \"list\" : [ { \"poa\" : \"poa\", \"resourceNm\" : \"resourceNm\", \"origin\" : \"origin\", \"entityId\" : \"entityId\", \"svcCd\" : \"svcCd\", \"updUsr\" : \"updUsr\", \"creDt\" : \"creDt\", \"rolNm\" : \"rolNm\", \"poaStatus\" : \"poaStatus\", \"creType\" : \"creType\", \"useYn\" : \"useYn\", \"creUsr\" : \"creUsr\", \"updDt\" : \"updDt\", \"contentType\" : \"contentType\", \"seq\" : \"seq\" }, { \"poa\" : \"poa\", \"resourceNm\" : \"resourceNm\", \"origin\" : \"origin\", \"entityId\" : \"entityId\", \"svcCd\" : \"svcCd\", \"updUsr\" : \"updUsr\", \"creDt\" : \"creDt\", \"rolNm\" : \"rolNm\", \"poaStatus\" : \"poaStatus\", \"creType\" : \"creType\", \"useYn\" : \"useYn\", \"creUsr\" : \"creUsr\", \"updDt\" : \"updDt\", \"contentType\" : \"contentType\", \"seq\" : \"seq\" } ], \"totalCount\" : 6 }, \"httpStatusCode\" : 0, \"status\" : \"status\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_RETRIEVE_SERVICE_SERVER = "/serviceServer/retrieveServiceServer";
    /**
     * POST /serviceServer/retrieveServiceServer : 서비스 서버 상세 조회
     * 서비스 서버 상세 조회
     *
     * @param retrieveServiceServerRequest  (required)
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "retrieveServiceServer",
        summary = "서비스 서버 상세 조회",
        description = "서비스 서버 상세 조회",
        tags = { "서비스 등록 - 서비스 서버 등록" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ComResponseSvcServer.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = ServiceServerApi.PATH_RETRIEVE_SERVICE_SERVER,
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    default ResponseEntity<ComResponseSvcServer> retrieveServiceServer(
        @Parameter(name = "RetrieveServiceServerRequest", description = "", required = true) @Valid @RequestBody RetrieveServiceServerRequest retrieveServiceServerRequest
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"code\" : \"code\", \"errorMessage\" : \"errorMessage\", \"detailMessage\" : \"detailMessage\", \"message\" : \"message\", \"body\" : { \"poa\" : \"192.168.1.0\", \"uplusYn\" : \"Y\", \"svcSvrType\" : \"svcSvrType\", \"svcCd\" : \"TEST\", \"updUsr\" : \"updUsr\", \"svcSvrEntityId\" : \"MN_CSE-S-51f72f5d94-NAMI\", \"uuid\" : \"51f72b7f-dd71-3373-8a82-b2b97d8f5d94\", \"creDt\" : \"creDt\", \"svcSvrCd\" : \"svcSvrCd\", \"svcSvrNum\" : 6, \"svcSvrNm\" : \"svcSvrNm\", \"rolNm\" : \"rolNm\", \"svcNm\" : \"테스트 서비스\", \"expireDuration\" : \"expireDuration\", \"useYn\" : \"useYn\", \"creUsr\" : \"creUsr\", \"updDt\" : \"updDt\", \"expireUnit\" : \"expireUnit\", \"contentType\" : \"json\" }, \"httpStatusCode\" : 0, \"status\" : \"status\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_RETRIEVE_SERVICE_SERVER_LIST = "/serviceServer/retrieveServiceServerList";
    /**
     * POST /serviceServer/retrieveServiceServerList : 서비스 서버 목록 조회
     * 서비스 서버 목록 조회
     *
     * @param retrieveServiceServerListRequest  (required)
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "retrieveServiceServerList",
        summary = "서비스 서버 목록 조회",
        description = "서비스 서버 목록 조회",
        tags = { "서비스 등록 - 서비스 서버 등록" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ComResponseSvcServerList.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = ServiceServerApi.PATH_RETRIEVE_SERVICE_SERVER_LIST,
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    default ResponseEntity<ComResponseSvcServerList> retrieveServiceServerList(
        @Parameter(name = "RetrieveServiceServerListRequest", description = "", required = true) @Valid @RequestBody RetrieveServiceServerListRequest retrieveServiceServerListRequest
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"code\" : \"code\", \"errorMessage\" : \"errorMessage\", \"detailMessage\" : \"detailMessage\", \"message\" : \"message\", \"body\" : { \"list\" : [ { \"poa\" : \"192.168.1.0\", \"uplusYn\" : \"Y\", \"svcSvrType\" : \"svcSvrType\", \"svcCd\" : \"TEST\", \"updUsr\" : \"updUsr\", \"svcSvrEntityId\" : \"MN_CSE-S-51f72f5d94-NAMI\", \"uuid\" : \"51f72b7f-dd71-3373-8a82-b2b97d8f5d94\", \"creDt\" : \"creDt\", \"svcSvrCd\" : \"svcSvrCd\", \"svcSvrNum\" : 6, \"svcSvrNm\" : \"svcSvrNm\", \"rolNm\" : \"rolNm\", \"svcNm\" : \"테스트 서비스\", \"expireDuration\" : \"expireDuration\", \"useYn\" : \"useYn\", \"creUsr\" : \"creUsr\", \"updDt\" : \"updDt\", \"expireUnit\" : \"expireUnit\", \"contentType\" : \"json\" }, { \"poa\" : \"192.168.1.0\", \"uplusYn\" : \"Y\", \"svcSvrType\" : \"svcSvrType\", \"svcCd\" : \"TEST\", \"updUsr\" : \"updUsr\", \"svcSvrEntityId\" : \"MN_CSE-S-51f72f5d94-NAMI\", \"uuid\" : \"51f72b7f-dd71-3373-8a82-b2b97d8f5d94\", \"creDt\" : \"creDt\", \"svcSvrCd\" : \"svcSvrCd\", \"svcSvrNum\" : 6, \"svcSvrNm\" : \"svcSvrNm\", \"rolNm\" : \"rolNm\", \"svcNm\" : \"테스트 서비스\", \"expireDuration\" : \"expireDuration\", \"useYn\" : \"useYn\", \"creUsr\" : \"creUsr\", \"updDt\" : \"updDt\", \"expireUnit\" : \"expireUnit\", \"contentType\" : \"json\" } ], \"totalCount\" : 6 }, \"httpStatusCode\" : 0, \"status\" : \"status\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_UPDATE_POA = "/serviceServer/poa/updatePoa";
    /**
     * POST /serviceServer/poa/updatePoa : POA 수정
     * POA 수정
     *
     * @param updatePoaRequest  (required)
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "updatePoa",
        summary = "POA 수정",
        description = "POA 수정",
        tags = { "서비스 서버 등록" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = ServiceServerApi.PATH_UPDATE_POA,
        consumes = { "application/json" }
    )
    default ResponseEntity<Void> updatePoa(
        @Parameter(name = "UpdatePoaRequest", description = "", required = true) @Valid @RequestBody UpdatePoaRequest updatePoaRequest
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_UPDATE_SERVICE_SERVER = "/serviceServer/updateServiceServer";
    /**
     * POST /serviceServer/updateServiceServer : 서비스 서버 수정
     * 서비스 서버 수정
     *
     * @param updateServiceServerRequest  (required)
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "updateServiceServer",
        summary = "서비스 서버 수정",
        description = "서비스 서버 수정",
        tags = { "서비스 등록 - 서비스 서버 등록" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = ServiceServerApi.PATH_UPDATE_SERVICE_SERVER,
        consumes = { "application/json" }
    )
    default ResponseEntity<Void> updateServiceServer(
        @Parameter(name = "UpdateServiceServerRequest", description = "", required = true) @Valid @RequestBody UpdateServiceServerRequest updateServiceServerRequest
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}